+++ tablet_qt/dialogs/logbox.h
@@ -38,6 +38,7 @@ public:
            bool offer_ok_at_end = true, int maximum_block_count = 1000,
            bool scroll_to_end_on_insert = true, bool word_wrap = false);
     ~LogBox() override;
+    void useWaitCursor(bool use_wait_cursor = true);
     void statusMessage(const QString& msg, bool as_html = false);
     void finish(bool success = true);
 public slots:
@@ -48,6 +49,7 @@ signals:
     void cancelled();
     void finished();
 protected:
+    bool m_use_wait_cursor;
     QPointer<QPlainTextEdit> m_editor;
     QPointer<QPushButton> m_ok;
     QPointer<QPushButton> m_cancel;